feat(task): update all task service

+++ src/main/java/com/br/scheduling/services/TaskService.java
@@ -3,9 +3,9 @@ package com.br.scheduling.services;
 import com.br.scheduling.dtos.TaskDTO;
 import com.br.scheduling.dtos.TaskDTOMapper;
 import com.br.scheduling.models.Task;
+import com.br.scheduling.models.enums.StatusTask;
 import com.br.scheduling.repositories.TaskRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,9 +19,15 @@ public class TaskService {
     private final TaskRepository taskRepository;
     private final TaskDTOMapper taskDTOMapper;
 
-    public Optional<Task> findTaskById(Long id){
-        //TODO
-        return taskRepository.findById(id);
+    public Optional<TaskDTO> findTaskById(Long id){
+        if (taskRepository.findById(id).isPresent()){
+            return taskRepository.findById(id)
+                    .map(taskDTOMapper);
+        }
+        else{
+            return null; //TODO better exception error message
+        }
+
     }
 
     public List<TaskDTO> getAllTasks(){
@@ -34,29 +40,48 @@ public class TaskService {
     public List<TaskDTO> pageSearchTask(){
         return null;
     }
-    public TaskDTO createTask(){
-        //TODO
-        return null;
+    public TaskDTO createTask(TaskDTO taskDTO){
+        Task task = new Task();
+
+        task.setName(taskDTO.name());
+        task.setPrice(taskDTO.price());
+        task.setStatusTask(StatusTask.AVAILABLE);
+
+        Task saveTask = taskRepository.save(task);
+
+        return taskDTOMapper.apply(saveTask);
     }
 
-    public TaskDTO updateTask(){
-        //TODO
-        return null;
+    public TaskDTO updateTask(TaskDTO taskDTO){
+        findTaskById(taskDTO.id());
+
+        Task task = new Task();
+
+        task.setName(taskDTO.name());
+        task.setPrice(taskDTO.price());
+        task.setStatusTask(taskDTO.statusTask());
+
+        Task updateTask = taskRepository.save(task);
+
+        return taskDTOMapper.apply(updateTask);
     }
 
-    public TaskDTO softDeleteTask(){
-        //TODO
-        return null;
+    public TaskDTO softDeleteTask(TaskDTO taskDTO){
+        findTaskById(taskDTO.id());
+
+        Task task = new Task();
+
+        task.setStatusTask(StatusTask.UNAVAILABLE);
+
+        Task updateTask = taskRepository.save(task);
+
+        return taskDTOMapper.apply(updateTask);
     }
 
     public void deleteTask(Long id){
-        //TODO verify if the task exist
-        taskRepository.deleteById(id);
+        if (findTaskById(id).isPresent()){
+            taskRepository.deleteById(id);
+        }
+        // TODO return if the id didn't exist
     }
-
-    public Optional<Task> verifyTaskExistById(Long id){
-        //TODO
-        return taskRepository.findById(id);
-    }
-
 }
